added a scatter plot of each pair of variables
For this code I got great help from the Seaborn Library, reference in the code and the Read me, seaborn has many features and simple lines of code for producing plots. It always amazes me when I research for a long time and find that what I want to do can be done with one or two lines of codes. Python is incredible that way.