added summary of each variable in a text file
I am not entirely happy with how this looks and so held back on uploading it, I will be uploading a better version when I figure out how to do it. I realised I have been making the mistake of doing most the editing on my own computer and not showing the process in the git commit history for the project.